+++ test/index.ts
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 console.log({ __dirname } as any);
 console.log({ __filename } as any);
 console.log({ ["import.meta.url"]: import.meta.url } as any);